+++ src/core/client.ts
@@ -1,5 +1,7 @@
 import { merge, mergeNotice, objectIsEmpty, makeNotice, makeBacktrace, runBeforeNotifyHandlers, newObject, logger, generateStackTrace, filter, filterUrl, formatCGIData } from './util'
-import { Config, Logger, BeforeNotifyHandler, AfterNotifyHandler, Notice } from './types'
+import {
+  Config, Logger, BreadcrumbRecord, BeforeNotifyHandler, AfterNotifyHandler, Notice, Noticeable
+} from './types'
 
 const notifier = {
   name: 'honeybadger-js',
@@ -20,12 +22,17 @@ const STRING_EMPTY = ''
 const NOT_BLANK = /\S/
 
 export default class Client {
+  /** @internal */
   private __pluginsExecuted = false
 
-  protected __context = {}
-  protected __breadcrumbs = []
-  protected __beforeNotifyHandlers = []
-  protected __afterNotifyHandlers = []
+  /** @internal */
+  protected __context: Record<string, unknown> = {}
+  /** @internal */
+  protected __breadcrumbs: BreadcrumbRecord[] = []
+  /** @internal */
+  protected __beforeNotifyHandlers: BeforeNotifyHandler[] = []
+  /** @internal */
+  protected __afterNotifyHandlers: AfterNotifyHandler[] = []
 
   config: Config
   logger: Logger
@@ -96,9 +103,9 @@ export default class Client {
     return this
   }
 
-  resetContext(context: Record<string, unknown>): Client {
+  resetContext(context?: Record<string, unknown>): Client {
     this.logger.warn('Deprecation warning: `Honeybadger.resetContext()` has been deprecated; please use `Honeybadger.clear()` instead.')
-    if (typeof context === 'object') {
+    if (typeof context === 'object' && context !== null) {
       this.__context = merge({}, context)
     } else {
       this.__context = {}
@@ -112,7 +119,7 @@ export default class Client {
     return this
   }
 
-  notify(notice: Partial<Notice>, name = undefined, extra = undefined): Record<string, unknown> | false | unknown {
+  notify(notice: Noticeable, name: string | Partial<Notice> = undefined, extra: Partial<Notice> = undefined): Record<string, unknown> | false | unknown {
     if (!this.config.apiKey) {
       this.logger.warn('Unable to send error report: no API key has been configured')
       return false
@@ -136,9 +143,9 @@ export default class Client {
     }
 
     if (name) {
-      notice = mergeNotice(notice, name)
+      notice = mergeNotice(notice, name as Partial<Notice>)
     }
-    if (typeof extra === 'object') {
+    if (typeof extra === 'object' && extra !== null) {
       notice = mergeNotice(notice, extra)
     }
 
@@ -186,7 +193,7 @@ export default class Client {
     return this.__send(notice)
   }
 
-  addBreadcrumb(message: string, opts: Record<string, unknown>): Client {
+  addBreadcrumb(message: string, opts?: Record<string, unknown>): Client {
     if (!this.config.breadcrumbsEnabled) { return }
 
     opts = opts || {}
@@ -196,9 +203,9 @@ export default class Client {
     const timestamp = new Date().toISOString()
 
     this.__breadcrumbs.push({
-      category: category,
+      category: category as string,
       message: message,
-      metadata: metadata,
+      metadata: metadata as Record<string, unknown>,
       timestamp: timestamp
     })
 
@@ -210,17 +217,20 @@ export default class Client {
     return this
   }
 
-  private __reportData(): boolean {
+  /** @internal */
+  protected __reportData(): boolean {
     if (this.config.reportData !== null) {
       return this.config.reportData
     }
     return !(this.config.environment && this.config.developmentEnvironments.includes(this.config.environment))
   }
 
-  protected __send(_notice: Partial<Notice>): unknown {
+  /** @internal */
+  protected __send(_notice: Partial<Notice>): boolean {
     throw (new Error('Must implement send in subclass'))
   }
 
+  /** @internal */
   protected __buildPayload(notice: Notice): Record<string, Record<string, unknown>> {
     const headers = filter(notice.headers, this.config.filters) || {}
     const cgiData = filter({
@@ -260,6 +270,7 @@ export default class Client {
     }
   }
 
+  /** @internal */
   protected __constructTags(tags: unknown): Array<string> {
     if (!tags) {
       return []