+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased][latest]
 ### Fixed
 - TypeScript: Add missing methods typedefs (#611)
+- Internal: Fixed TypeScript types, added null checks, automated type declaration files
 
 ## [3.2.1] - 2021-05-17
 ### Fixed